Adding `react-js-diagrams` and some refactors
At the `react-js-diagrams` repository there's a bunch of cool
examples. I took the liberty to read them and extract as much
information from there that I could.

I also did some code extraction (see `diagram.helper.js` file)
and some refactoring, not only to make it more organized but
learn a bit more about the library as well.

It seems to be working up to this point... :)