Adding functional DnD and nodes panel
Geez... that was troublesome:
Firstly, while analyzing the code of `react-js-diagrams` demo 4,
which cover almost all specs that I need, I saw the usage of Redux
there quite unnecessary.

So, with that in mind, I took the liberty to not waste time thinking
about the nodes and the Drag-n-Drop (the latter would certainly take
some time due to its complexity) and focus on removing the Redux
flow from that example. Not only because it was unnecessary, but
to understand the data flow.

Ironically, it took me HOURS and I got bunch of `Can't assign to read-
only property` errors, which (what is even more ironic) made me
understand a lot more about Redux and why it was used in this context.

Now now... back to what it matters:
- Diagram: ✔
- Nodes: ✔
- Nodes Panel: ✔
- Drag-n-drop nodes to diagram: ✔
- Local storage persistence: [pending]
- (optional) Custom shape nodes: [pending]